pyproject.toml
[project]
name = "tdd-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "aiosqlite>=0.21.0",
    "asyncio>=4.0.0",
    "greenlet>=3.2.4",
    "litestar>=2.18.0",
    "ptbtest>=1.3",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    "yookassa>=3.8.0",
]
